+++ CHANGELOG.md
@@ -2,7 +2,15 @@
 
 *Changelog created using the [Simple Changelog](https://marketplace.visualstudio.com/items?itemName=tobiaswaelde.vscode-simple-changelog) extension for VS Code.*
 
-## [0.1.0] - 2023-11-26
+## [1.2.0] - 2023-12-10
+
+### Added
+
+- new main.py file. Debugged previous version
+- python.code-snippets added
+- config file can now be used to generate code
+
+## [1.1.0] - 2023-11-26
 
 ### Added